+++ DP5/lib/Carbon_assignment.py
