+++ DP5/lib/Proton_assignment.py
